Update taskscreen (#8)
* UI updated. Displaying more task data

* Task screen UI update. Delete option added on screen.

* Task screen - trying out some aesthetic changes

* TaskScreen UI update. Confirm delete alert added.

* Added alert dialogs

* User can now remove location from reminder. Also, some refactoring

* Preference screen - proximity distance setting now can be set to the tenth decimal

* Applying new font

* Random code edits. Also switched to explicit broadcast calls due to Android O

* switching to proximity alert management

+++ app/src/main/java/com/nghianguyen/intheneighborhood/ui/tasklist/adapter/TaskListItemPresenter.java
@@ -35,7 +35,7 @@ public class TaskListItemPresenter {
         recyclerView.showContextMenuForChild(view);
     }
 
-    public void updateNearbyTasks(Location location, int proximityDistance){
+    public void updateNearbyTasks(Location location, float proximityDistance){
         List<Task> nearbyTasks = new ArrayList<>();
 
         LatLng locLatLng;
